exclude icon-theme.cache
swupd logs continuosly list icon-theme.cache files as missing,
giving the wrong impression that they're not appropriately created,
due to a post-update task that re-creates them in a different location.

This commit cleans the swupd logs by ommiting providing the
cache files, which are now created and updated by a swupd trigger
service.

Signed-off-by: Jose Lamego <jose.a.lamego@intel.com>